Logist Test Fix

+++ src/main/java/com/diploma/maksimov/restcontroller/LogistController.java
@@ -1,6 +1,7 @@
 package com.diploma.maksimov.restcontroller;
 
 import com.diploma.maksimov.dto.Logist;
+import com.diploma.maksimov.service.EmployeeService;
 import com.diploma.maksimov.service.ILogistService;
 import com.diploma.maksimov.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/rest/boss/employee/{id}")
+@RequestMapping("/rest/boss/employee")
 public class LogistController {
     private final ILogistService logistService;
 
@@ -23,49 +24,64 @@ public class LogistController {
     @Autowired
     private UserService userService;
 
-    @PostMapping(value = "/logist")
+    @Autowired
+    private EmployeeService employeeService;
+
+    @PostMapping(value = "/{id}/logist")
     public ResponseEntity<?> create(@RequestBody Logist logist, @PathVariable Long id) {
-        logistService.create(logist);
-        userService.addRole(id,4L);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+        if (employeeService.read(id) != null) {
+            logistService.create(logist);
+            userService.addRole(id,4L);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-//    @GetMapping(value = "/logist")
-//    public ResponseEntity<List<Logist>> readAll() {
-//        final List<Logist> logists = logistService.readAll();
-//
-//        if (logists != null && !logists.isEmpty()) {
-//            return new ResponseEntity<>(logists, HttpStatus.OK);
-//        }
-//        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//    }
-
     @GetMapping(value = "/logist")
-    public ResponseEntity<Logist> read(@PathVariable long id) {
-        final Logist logist = logistService.read(id);
+    public ResponseEntity<List<Logist>> readAll() {
+        final List<Logist> logists = logistService.readAll();
 
-        if (logist != null) {
-            return new ResponseEntity<>(logist, HttpStatus.OK);
+        if (logists != null && !logists.isEmpty()) {
+            return new ResponseEntity<>(logists, HttpStatus.OK);
         }
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping(value = "/logist")
-    public ResponseEntity<?> update(@PathVariable long id, @RequestBody Logist logist) {
-        final boolean updated = logistService.update(logist, id);
-        if (updated) {
-            return new ResponseEntity<>(HttpStatus.OK);
+    @GetMapping(value = "/{id}/logist")
+    public ResponseEntity<Logist> read(@PathVariable long id) {
+        if (employeeService.read(id) != null) {
+            final Logist logist = logistService.read(id);
+
+            if (logist != null) {
+                return new ResponseEntity<>(logist, HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @DeleteMapping(value = "/logist")
-    public ResponseEntity<?> delete(@PathVariable long id) {
-        final boolean deleted = logistService.delete(id);
-        userService.deleteRole(id,4L);
-        if (deleted) {
-            return new ResponseEntity<>(HttpStatus.OK);
+    @PutMapping(value = "/{id}/logist")
+    public ResponseEntity<?> update(@PathVariable long id, @RequestBody Logist logist) {
+        if (employeeService.read(id) != null) {
+            final boolean updated = logistService.update(logist, id);
+            if (updated) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @DeleteMapping(value = "/{id}/logist")
+    public ResponseEntity<?> delete(@PathVariable long id) {
+        if (employeeService.read(id) != null) {
+            final boolean deleted = logistService.delete(id);
+            userService.deleteRole(id,4L);
+            if (deleted) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 }
